:recycle: 테스트 메인 페이지 컴포넌트로 분리
